Refactor pypesa and decouple __init__.py
